requirements.txt
crewai[tools]>=1.0.0,<2.0.0
langchain>=0.2.6,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-community>=0.2.6,<0.4.0
openai>=1.40.0,<3.0.0
pydantic>=2.10,<3.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
requests-oauthlib>=1.3.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
youtube-transcript-api>=1.0.0,<2.0.0
PyYAML>=6.0,<7.0
pytest>=7.0.0,<9.0.0
pytest-mock>=3.10.0,<4.0.0
